+++ src/calculationsWithFractions/Calculator.java
@@ -30,13 +30,13 @@ public class Calculator {
 
 	public Fraction add() {
 		bringToLowestCommonDenominator();
-		int numerator = fraction1.getNumerator() + fraction2.getNumerator();
+		long numerator = fraction1.getNumerator() + fraction2.getNumerator();
 		return new Fraction(numerator, fraction1.getDenominator());
 	}
 	
 	public Fraction subtract() {
 		bringToLowestCommonDenominator();
-		int numerator = fraction1.getNumerator() - fraction2.getNumerator();
+		long numerator = fraction1.getNumerator() - fraction2.getNumerator();
 		return new Fraction(numerator, fraction1.getDenominator());
 	}
 	
@@ -45,8 +45,8 @@ public class Calculator {
 		if (reciprocal.length == 1) {
 			multiplyer = reciprocal[0];
 		}
-		int numerator = fraction1Original.getNumerator() * multiplyer.getNumerator();
-		int denominator = fraction1Original.getDenominator() * multiplyer.getDenominator();
+		long numerator = fraction1Original.getNumerator() * multiplyer.getNumerator();
+		long denominator = fraction1Original.getDenominator() * multiplyer.getDenominator();
 		return new Fraction(numerator, denominator);
 	}
 	
@@ -56,12 +56,12 @@ public class Calculator {
 	}
 	
 	private void bringToLowestCommonDenominator() {
-		int denominator1 = fraction1.getDenominator();
-		int denominator2 = fraction2.getDenominator();
+		long denominator1 = fraction1.getDenominator();
+		long denominator2 = fraction2.getDenominator();
 		if (denominator1 != denominator2) {
-			int higherDenominator = Math.max(denominator1, denominator2);
-			int lowerDenominator = Math.min(denominator1, denominator2);
-			int commonDenominator = higherDenominator;
+			long higherDenominator = Math.max(denominator1, denominator2);
+			long lowerDenominator = Math.min(denominator1, denominator2);
+			long commonDenominator = higherDenominator;
 			while (commonDenominator % lowerDenominator != 0) {
 				commonDenominator += higherDenominator;
 			}
